перевести проект на jpa (hibernate) #17

+++ src/test/resources/h2/init.sql
@@ -16,16 +16,16 @@ CREATE TABLE profession
 
 CREATE TABLE player
 (
-    id             SERIAL PRIMARY KEY,
-    name           VARCHAR(12) NOT NULL,
-    title          VARCHAR(30) NULL,
-    race_id        INTEGER REFERENCES race(id),
-    experience     INTEGER     CHECK ( 0 < experience AND experience <= 10000000 ),
-    profession_id     INTEGER REFERENCES profession(id),
-    level          INTEGER check ( 0 <= level ) NOT NULL,
+    id               BIGSERIAL PRIMARY KEY,
+    name             VARCHAR(12) NOT NULL,
+    title            VARCHAR(30) NULL,
+    race_id          INTEGER REFERENCES race(id),
+    experience       INTEGER     CHECK ( 0 < experience AND experience <= 10000000 ),
+    profession_id    INTEGER REFERENCES profession(id),
+    level            INTEGER check ( 0 <= level ) NOT NULL,
     until_next_level INTEGER NOT NULL,
-    birthday       TIMESTAMP NOT NULL,
-    banned         BOOLEAN NOT NULL
+    birthday         TIMESTAMP NOT NULL,
+    banned           BOOLEAN NOT NULL
 );
 
 INSERT INTO race(name)
